requirements.txt
pyautogui>=0.9.54
pillow>=10.0.0
pynput>=1.7.6
python-dotenv>=1.0.0
google-generativeai>=0.3.2
requests>=2.31.0
pandas>=2.2.0
matplotlib>=3.8.0
psutil>=5.9.0